fix(layerzero-onft): config import causes running connect-contracts script
