+++ layerzero-onft/deploy/connect-contracts.ts
@@ -1,20 +1,8 @@
 import { Contract, JsonRpcProvider, Wallet, zeroPadValue } from "ethers";
-import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { vars } from "hardhat/config";
+import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { Provider, Wallet as ZkWallet } from "zksync-ethers";
-
-export const config = {
-  baseSepolia: {
-    endpointId: 40245,
-    // Replace with your own NFT contract address
-    nftContractAddress: "0x64D71429d50172CD3B759126C4C43a3dA8B0a038",
-  },
-  abstractTestnet: {
-    endpointId: 40313,
-    // Replace with your own NFT contract address
-    nftContractAddress: "0x8442F3845C460AA33eDB2837A9daf627C4D0EfF2",
-  },
-};
+import { config } from "./config";
 
 export default async function connectContract(hre: HardhatRuntimeEnvironment) {
   console.log(`Running connect-contracts script`);
